pyproject.toml
[project]
name = "your-cross"
version = "0.1.0"
description = ""
authors = [
    {name = "cileno",email = "cileno12@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12, <4.0"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "streamlit (>=1.45.1,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]

[tool.poetry]
packages = [{include = "your_cross", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-django = "^4.11.1"
coverage = "^7.8.2"
pytest-cov = "^6.1.1"
blue = "^0.9.1"
isort = "^6.0.1"
taskipy = "^1.14.1"


[tool.poetry.group.doc.dependencies]
mkdocs = "^1.6.1"
mkdocs-material = "^9.6.14"
mkdocstrings-python = "^1.16.11"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "."
addopts = "--doctest-modules"


[tool.black]
line-length = 79
target-version = ["py312"]

[tool.isort]
profile = "black"

[tool.taskipy.tasks]
runserver = "python your_cross/manage.py runserver"
streamlit = "streamlit run streamlit_app/Home.py"
lint = "blue --check --diff . && isort --check --diff ."
format = "black . && blue . && isort ."
docs = "mkdocs serve"
pre_test = "task lint"
test = "pytest -s -x --cov=your_cross -vv"
post_test = "coverage html"
